feat: centralize workout data queries
