+++ src/hooks/useWorkoutHistory.ts
@@ -1,6 +1,7 @@
 
 import { useState, useEffect, useCallback } from 'react';
 import { supabase } from '@/integrations/supabase/client';
+import { getWorkoutHistory } from '@/services/workout/workoutDataService';
 import { useQuery, useQueryClient } from '@tanstack/react-query';
 
 // Define WorkoutHistoryFilters interface to use throughout the application
@@ -57,86 +58,7 @@ export function useWorkoutHistory(filters: WorkoutHistoryFilters = { limit: 30 }
   
   const fetchWorkouts = useCallback(async () => {
     try {
-      // Build the count query first to get total number of workouts
-      const countQuery = supabase
-        .from('workout_sessions')
-        .select('id', { count: 'exact', head: true });
-        
-      // Apply filters to count query
-      if (filters.startDate) {
-        countQuery.gte('start_time', filters.startDate);
-      }
-      
-      if (filters.endDate) {
-        countQuery.lte('start_time', filters.endDate);
-      }
-      
-      if (filters.trainingTypes && filters.trainingTypes.length > 0) {
-        countQuery.in('training_type', filters.trainingTypes);
-      }
-      
-      const { count, error: countError } = await countQuery;
-      
-      if (countError) throw countError;
-      
-      // Now build the main query
-      let query = supabase
-        .from('workout_sessions')
-        .select('*')
-        .order('start_time', { ascending: false });
-        
-      if (filters.limit) {
-        query = query.limit(filters.limit);
-      }
-      
-      if (filters.offset) {
-        query = query.range(filters.offset, filters.offset + (filters.limit || 30) - 1);
-      }
-      
-      if (filters.startDate) {
-        query = query.gte('start_time', filters.startDate);
-      }
-      
-      if (filters.endDate) {
-        query = query.lte('start_time', filters.endDate);
-      }
-      
-      if (filters.trainingTypes && filters.trainingTypes.length > 0) {
-        query = query.in('training_type', filters.trainingTypes);
-      }
-      
-      const { data, error: workoutsError } = await query;
-      
-      if (workoutsError) throw workoutsError;
-      
-      // For each workout, get the exercise count and set count
-      const exerciseCountData: Record<string, { exercises: number; sets: number }> = {};
-      
-      await Promise.all(
-        (data || []).map(async (workout) => {
-          const { data: exerciseSets, error: exerciseSetsError } = await supabase
-            .from('exercise_sets')
-            .select('exercise_name, id')
-            .eq('workout_id', workout.id);
-            
-          if (exerciseSetsError) throw exerciseSetsError;
-          
-          // Count unique exercise names
-          const exerciseNames = new Set();
-          exerciseSets?.forEach(set => exerciseNames.add(set.exercise_name));
-          
-          exerciseCountData[workout.id] = {
-            exercises: exerciseNames.size,
-            sets: exerciseSets?.length || 0
-          };
-        })
-      );
-      
-      return {
-        workouts: data || [],
-        exerciseCounts: exerciseCountData,
-        totalCount: count || 0
-      };
+      return await getWorkoutHistory(filters);
     } catch (err) {
       console.error('Error fetching workout history:', err);
       throw err;
